LPS-92606 Rename
