+++ modules/apps/asset/asset-list-web/src/main/java/com/liferay/asset/list/web/internal/servlet/taglib/ui/AssetListFilterFormNavigatorEntry.java
@@ -29,8 +29,8 @@ import javax.servlet.ServletContext;
 	property = "form.navigator.entry.order:Integer=400",
 	service = FormNavigatorEntry.class
 )
-public class FilterFormNavigatorEntry
-	extends BaseConfigurationFormNavigatorEntry {
+public class AssetListFilterFormNavigatorEntry
+	extends BaseAssetListFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
