+++ modules/apps/asset/asset-list-web/src/main/java/com/liferay/asset/list/web/internal/servlet/taglib/ui/AssetListOrderingFormNavigatorEntry.java
@@ -15,9 +15,11 @@
 package com.liferay.asset.list.web.internal.servlet.taglib.ui;
 
 import com.liferay.asset.publisher.constants.AssetPublisherConstants;
-import com.liferay.asset.publisher.constants.AssetPublisherPortletKeys;
+import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizer;
+import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizerRegistry;
+import com.liferay.portal.kernel.model.Portlet;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.portlet.PortletIdCodec;
+import com.liferay.portal.kernel.service.PortletLocalService;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
 import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
@@ -32,11 +34,11 @@ import javax.servlet.ServletContext;
  * @author Eudaldo Alonso
  */
 @Component(
-	property = "form.navigator.entry.order:Integer=400",
+	property = "form.navigator.entry.order:Integer=200",
 	service = FormNavigatorEntry.class
 )
-public class ScopeFormNavigatorEntry
-	extends BaseConfigurationFormNavigatorEntry {
+public class AssetListOrderingFormNavigatorEntry
+	extends BaseAssetListFormNavigatorEntry {
 
 	@Override
 	public String getCategoryKey() {
@@ -45,17 +47,13 @@ public class ScopeFormNavigatorEntry
 
 	@Override
 	public String getKey() {
-		return "scope";
+		return "ordering";
 	}
 
 	@Override
 	public boolean isVisible(User user, Object object) {
-		if (!isManualSelection() && !isDynamicAssetSelection()) {
+		if (!isDynamicAssetSelection()) {
 			return false;
-		}
-
-		if (isManualSelection()) {
-			return true;
 		}
 
 		ServiceContext serviceContext =
@@ -65,14 +63,19 @@ public class ScopeFormNavigatorEntry
 
 		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
 
-		String rootPortletId = PortletIdCodec.decodePortletName(
-			portletDisplay.getPortletName());
+		Portlet portlet = _portletLocalService.getPortletById(
+			themeDisplay.getCompanyId(), portletDisplay.getPortletResource());
 
-		if (rootPortletId.equals(AssetPublisherPortletKeys.RELATED_ASSETS)) {
-			return false;
+		AssetPublisherCustomizer assetPublisherCustomizer =
+			assetPublisherCustomizerRegistry.getAssetPublisherCustomizer(
+				portlet.getRootPortletId());
+
+		if (assetPublisherCustomizer == null) {
+			return true;
 		}
 
-		return true;
+		return assetPublisherCustomizer.isOrderingAndGroupingEnabled(
+			serviceContext.getRequest());
 	}
 
 	@Override
@@ -86,7 +89,13 @@ public class ScopeFormNavigatorEntry
 
 	@Override
 	protected String getJspPath() {
-		return "/configuration/scope.jsp";
+		return "/configuration/ordering.jsp";
 	}
+
+	@Reference
+	protected AssetPublisherCustomizerRegistry assetPublisherCustomizerRegistry;
+
+	@Reference
+	private PortletLocalService _portletLocalService;
 
 }
