+++ modules/apps/asset/asset-list-web/src/main/java/com/liferay/asset/list/web/internal/servlet/taglib/ui/AssetListSourceFormNavigatorEntry.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.asset.list.web.internal.servlet.taglib.ui;
+
+import com.liferay.asset.publisher.constants.AssetPublisherConstants;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+import javax.servlet.ServletContext;
+
+/**
+ * @author Eudaldo Alonso
+ */
+@Component(
+	property = "form.navigator.entry.order:Integer=500",
+	service = FormNavigatorEntry.class
+)
+public class AssetListSourceFormNavigatorEntry
+	extends BaseAssetListFormNavigatorEntry {
+
+	@Override
+	public String getCategoryKey() {
+		return AssetPublisherConstants.CATEGORY_KEY_ASSET_SELECTION;
+	}
+
+	@Override
+	public String getKey() {
+		return "source";
+	}
+
+	@Override
+	public boolean isVisible(User user, Object object) {
+		return isDynamicAssetSelection();
+	}
+
+	@Override
+	@Reference(
+		target = "(osgi.web.symbolicname=com.liferay.asset.publisher.web)",
+		unbind = "-"
+	)
+	public void setServletContext(ServletContext servletContext) {
+		super.setServletContext(servletContext);
+	}
+
+	@Override
+	protected String getJspPath() {
+		return "/configuration/source.jsp";
+	}
+
+}
