+++ modules/apps/asset/asset-list-web/src/main/java/com/liferay/asset/list/web/internal/servlet/taglib/ui/BaseAssetListFormNavigatorEntry.java
@@ -31,7 +31,7 @@ import java.util.Objects;
 /**
  * @author Eudaldo Alonso
  */
-public abstract class BaseConfigurationFormNavigatorEntry
+public abstract class BaseAssetListFormNavigatorEntry
 	extends BaseJSPFormNavigatorEntry {
 
 	@Override
