--- modules/apps/asset/asset-list-web/src/main/java/com/liferay/asset/list/web/internal/servlet/taglib/ui/OrderingFormNavigatorEntry.java
@@ -1,101 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.asset.list.web.internal.servlet.taglib.ui;
-
-import com.liferay.asset.publisher.constants.AssetPublisherConstants;
-import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizer;
-import com.liferay.asset.publisher.web.internal.util.AssetPublisherCustomizerRegistry;
-import com.liferay.portal.kernel.model.Portlet;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.service.PortletLocalService;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.ServiceContextThreadLocal;
-import com.liferay.portal.kernel.servlet.taglib.ui.FormNavigatorEntry;
-import com.liferay.portal.kernel.theme.PortletDisplay;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-import javax.servlet.ServletContext;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(
-	property = "form.navigator.entry.order:Integer=200",
-	service = FormNavigatorEntry.class
-)
-public class OrderingFormNavigatorEntry
-	extends BaseConfigurationFormNavigatorEntry {
-
-	@Override
-	public String getCategoryKey() {
-		return AssetPublisherConstants.CATEGORY_KEY_ASSET_SELECTION;
-	}
-
-	@Override
-	public String getKey() {
-		return "ordering";
-	}
-
-	@Override
-	public boolean isVisible(User user, Object object) {
-		if (!isDynamicAssetSelection()) {
-			return false;
-		}
-
-		ServiceContext serviceContext =
-			ServiceContextThreadLocal.getServiceContext();
-
-		ThemeDisplay themeDisplay = serviceContext.getThemeDisplay();
-
-		PortletDisplay portletDisplay = themeDisplay.getPortletDisplay();
-
-		Portlet portlet = _portletLocalService.getPortletById(
-			themeDisplay.getCompanyId(), portletDisplay.getPortletResource());
-
-		AssetPublisherCustomizer assetPublisherCustomizer =
-			assetPublisherCustomizerRegistry.getAssetPublisherCustomizer(
-				portlet.getRootPortletId());
-
-		if (assetPublisherCustomizer == null) {
-			return true;
-		}
-
-		return assetPublisherCustomizer.isOrderingAndGroupingEnabled(
-			serviceContext.getRequest());
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.asset.publisher.web)",
-		unbind = "-"
-	)
-	public void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
-	}
-
-	@Override
-	protected String getJspPath() {
-		return "/configuration/ordering.jsp";
-	}
-
-	@Reference
-	protected AssetPublisherCustomizerRegistry assetPublisherCustomizerRegistry;
-
-	@Reference
-	private PortletLocalService _portletLocalService;
-
-}
